fix(release): remove transient release plan
