+++ .github/actions/release/action.yaml
@@ -17,4 +17,4 @@ runs:
       run: |
         source .env
         export NODE_OPTIONS
-        yarn release plan foreach -v --exclude . release create
+        yarn workspace @atls/raijin release create
