+++ .github/actions/version/action.yml
@@ -5,17 +5,28 @@ inputs:
   GITHUB_TOKEN:
     required: true
 
+outputs:
+  release_required:
+    description: Whether merged changes produced a Raijin deferred version record
+    value: ${{ steps.version.outputs.release_required }}
+
 runs:
   using: composite
 
   steps:
     - name: Apply deferred versions
+      id: version
       shell: bash
       run: |
         source .env
         export NODE_OPTIONS
         yarn release version defer
-        yarn release plan create --output .raijin/release-plan.json
+        if [ -z "$(git status --porcelain -- .yarn/versions)" ]; then
+          echo "release_required=false" >> "$GITHUB_OUTPUT"
+          echo "No Raijin release target"
+          exit 0
+        fi
+        echo "release_required=true" >> "$GITHUB_OUTPUT"
         yarn version apply --all
       env:
         GITHUB_TOKEN: ${{ inputs.GITHUB_TOKEN }}
