+++ .github/workflows/publish.yaml
@@ -61,6 +61,7 @@ jobs:
           SELECTED_WORKSPACE: ${{ inputs.workspace }}
 
       - name: Version
+        id: version
         if: ${{ github.event.pull_request.merged == true }}
         uses: ./.github/actions/version
         with:
@@ -89,7 +90,7 @@ jobs:
           YARN_NPM_PUBLISH_REGISTRY: https://npm.pkg.github.com
 
       - name: Publish to NPM
-        if: ${{ success() && (github.event_name == 'workflow_dispatch' || github.event.pull_request.merged == true) }}
+        if: ${{ success() && (github.event_name == 'workflow_dispatch' || (github.event.pull_request.merged == true && steps.version.outputs.release_required == 'true')) }}
         id: npm
         uses: ./.github/actions/publish-npm
         with:
@@ -100,7 +101,7 @@ jobs:
           YARN_NPM_PUBLISH_REGISTRY: https://registry.npmjs.org
 
       - name: Publish to GitHub Packages
-        if: ${{ success() && (github.event_name == 'workflow_dispatch' || github.event.pull_request.merged == true) }}
+        if: ${{ success() && (github.event_name == 'workflow_dispatch' || (github.event.pull_request.merged == true && steps.version.outputs.release_required == 'true')) }}
         id: github-packages
         uses: ./.github/actions/publish-github-packages
         with:
@@ -111,17 +112,17 @@ jobs:
           YARN_NPM_PUBLISH_REGISTRY: https://npm.pkg.github.com
 
       - name: Bundle yarn
-        if: ${{ success() && github.event.pull_request.merged == true }}
+        if: ${{ success() && github.event.pull_request.merged == true && steps.version.outputs.release_required == 'true' }}
         uses: ./.github/actions/bundle
 
       - name: Create release
-        if: ${{ success() && github.event.pull_request.merged == true }}
+        if: ${{ success() && github.event.pull_request.merged == true && steps.version.outputs.release_required == 'true' }}
         uses: ./.github/actions/release
         with:
           GITHUB_TOKEN: ${{ steps.app-token.outputs.token }}
 
       - name: Commit changes
-        if: ${{ success() && github.event.pull_request.merged == true }}
+        if: ${{ success() && github.event.pull_request.merged == true && steps.version.outputs.release_required == 'true' }}
         uses: ./.github/actions/commit
         with:
           GITHUB_TOKEN: ${{ steps.app-token.outputs.token }}
