+++ docs/raijin/README.ru.md
@@ -37,6 +37,6 @@
 
 ## Покрытие текущей версии
 
-- Команд: 37 (active: 36, inactive: 1)
+- Команд: 35 (active: 34, inactive: 1)
 - Workspace-пакетов: 59
-- Последняя генерация: 2026-07-01T18:13:03.732Z
+- Последняя генерация: 2026-07-01T21:53:56.457Z
