+++ docs/raijin/index.meta.v1.json
@@ -1,11 +1,11 @@
 {
   "schemaVersion": 1,
   "generatedBy": "scripts/raijin/generate-artifacts.mjs",
-  "contentSha256": "2b1cb903a3f2cf987ef0440169bc895afca42633ae98df905a6c85ae49fa33a2",
+  "contentSha256": "330304bbd9f6d07100d1207a0c702bc3b50b4c66825c1bd5ed6b67b5908615fd",
   "packageManager": "yarn@4.14.1",
   "workspaceCount": 59,
-  "commandCount": 37,
-  "activeCommandCount": 36,
+  "commandCount": 35,
+  "activeCommandCount": 34,
   "inactiveCommandCount": 1,
-  "lastGenerated": "2026-07-01T18:13:03.732Z"
+  "lastGenerated": "2026-07-01T21:53:56.457Z"
 }
